+++ src/textual/widgets/tree.py
@@ -1,5 +1,19 @@
 """Make non-widget Tree support classes available."""
 
-from ._tree import TreeNode
+from ._tree import (
+    EventTreeDataType,
+    NodeID,
+    RemoveRootError,
+    TreeDataType,
+    TreeNode,
+    UnknownNodeID,
+)
 
-__all__ = ["TreeNode"]
+__all__ = [
+    "EventTreeDataType",
+    "NodeID",
+    "RemoveRootError",
+    "TreeDataType",
+    "TreeNode",
+    "UnknownNodeID",
+]